% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/116_summarize_DB.R
\name{add_DB_subset}
\alias{add_DB_subset}
\title{add_DB_subset}
\usage{
add_DB_subset(
  DB,
  subset_name,
  filter_field,
  filter_choices,
  dir_other = file.path(DB$dir_path, "output"),
  file_name = paste0("PSDB_", subset_name),
  form_names = NULL,
  field_names = NULL,
  deidentify = T,
  force = F
)
}
\description{
add_DB_subset
}
