% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_transform_DB.R
\name{add_default_forms_transformation}
\alias{add_default_forms_transformation}
\title{add_default_forms_transformation}
\usage{
add_default_forms_transformation(DB)
}
\value{
DB object
}
\description{
add_default_forms_transformation
}
