% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_new_vars.R
\name{add_edit_fields}
\alias{add_edit_fields}
\title{add_edit_fields}
\usage{
add_edit_fields(
  DB,
  field_name,
  form_name,
  field_type,
  field_type_R = NA,
  field_label = NA,
  select_choices_or_calculations = NA,
  field_note = NA,
  identifier = "",
  units = NA,
  insert_after,
  data_func = NULL
)
}
\description{
add_edit_fields
}
