% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_transform_DB.R
\name{add_field_transformation}
\alias{add_field_transformation}
\title{add_field_transformation}
\usage{
add_field_transformation(
  DB,
  field_name,
  form_name,
  field_type,
  field_type_R = NA,
  field_label = NA,
  select_choices_or_calculations = NA,
  field_note = NA,
  identifier = "",
  units = NA,
  data_func = NA
)
}
\value{
DB object
}
\description{
add_field_transformation
}
