% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapSync-experimental.R
\name{add_fields_to_data_list}
\alias{add_fields_to_data_list}
\title{add_fields_to_data_list}
\usage{
add_fields_to_data_list(data_list, transformation)
}
\description{
add_fields_to_data_list
}
