% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/111_process.R
\name{add_ID_to_DF}
\alias{add_ID_to_DF}
\title{add REDCap ID to any dataframe using a ref_id}
\usage{
add_ID_to_DF(DF, DB, ref_id)
}
\arguments{
\item{DF}{dataframe}

\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{ref_id}{column name that matches a REDCap variable name that could be an ALT id such as MRN}
}
\value{
original dataframe with REDCap id_col added as the first column
}
\description{
add REDCap ID to any dataframe using a ref_id
}
