% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_new_vars.R
\name{add_new_varriable}
\alias{add_new_varriable}
\title{add_new_varriable}
\usage{
add_new_varriable(
  DB,
  field_name,
  form_name,
  field_type,
  field_type_R = NA,
  field_label = NA,
  select_choices_or_calculations = NA,
  field_note = NA,
  identifier = "",
  units = NA,
  insert_after,
  data_func = NULL
)
}
\description{
add_new_varriable
}
