% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_all_functions.R
\name{blank_DB}
\alias{blank_DB}
\title{blank DB object}
\usage{
blank_DB()

blank_DB()
}
\value{
blank_DB list for reference

blank_DB list for reference
}
\description{
blank DB object

blank DB object
}
