% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_setup.R
\name{cache_clear}
\alias{cache_clear}
\title{Clear your cached projects}
\usage{
cache_clear()
}
\value{
message confirmation

The file path of your RosyREDCap cache

your path
}
\description{
Included for transparency and confirmation/testing.
}
\details{
This function checks the location of the cache established by [hoardr::hoard()].
}
\examples{
\dontrun{
path <- cache_path()
print(path)
}
}
\seealso{
Other cache: 
\code{\link{cache_path}()}
}
\concept{Cache}
\concept{cache}
\keyword{cache,}
\keyword{internal}
