% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_all_functions.R
\name{check_redcap_log}
\alias{check_redcap_log}
\title{Check the REDCap log}
\usage{
check_redcap_log(DB, last = 24, units = "hours", begin_time = "")
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{last}{numeric paired with units. Default is 24.}

\item{units}{character paired with last. Options are "mins","hours","days". Default is "hours".}

\item{begin_time}{character of time where the log should start from. Example 2023-07-11 13:15:06.}
}
\value{
data.frame of log that has been cleaned and has extra summary columns
}
\description{
Check the REDCap log
}
