% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_annotate_DB.R
\name{clean_column_for_table}
\alias{clean_column_for_table}
\title{clean column for plotting; manual addition of clean_DB}
\usage{
clean_column_for_table(col, class, label, units, levels)
}
\arguments{
\item{col}{the column vector}

\item{class}{character for column type: integer, factor, numeric}

\item{label}{character for label}

\item{units}{character for units}

\item{levels}{character vector of levels for factor}
}
\value{
cleaned column
}
\description{
Turns choices into factors and integers to integer for table processing such as with table1 and plots
}
