% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_redcap_process.R
\name{clean_DB}
\alias{clean_DB}
\title{clean DB columns for plotting using the metadata}
\usage{
clean_DB(
  DB,
  drop_blanks = T,
  drop_unknowns = T,
  units_df,
  merge_non_repeating = F
)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{drop_blanks}{logical for dropping n=0 choices}

\item{drop_unknowns}{logical for dropping missing codes}

\item{units_df}{data.frame with two columns: `field_name` in the metadata and `units` to set units}
}
\value{
DB object cleaned for table or plots
}
\description{
Turns choices into factors and integers to integer for table processing such as with table1 and plots
}
