% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_redcap_process.R
\name{clean_to_raw_form}
\alias{clean_to_raw_form}
\title{Clean to Raw REDCap forms}
\usage{
clean_to_raw_form(FORM, DB)
}
\arguments{
\item{FORM}{data.frame of clean REDCap to be converted to raw REDCap (for uploads)}

\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
DB object that has been filtered to only include the specified records
}
\description{
Clean to Raw REDCap forms
}
