% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapSync-experimental.R
\name{clear_project_fields}
\alias{clear_project_fields}
\title{clear_project_fields}
\usage{
clear_project_fields(project)
}
\description{
`r lifecycle::badge("experimental")`
Clears and project fields that were added with `add_project_fields()`
or if `setup_project()` contained `add_default_fields = TRUE`
}
