% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyUtils.R
\name{combine_R_files}
\alias{combine_R_files}
\title{combine_R_files}
\usage{
combine_R_files(
  source_dir = file.path(getwd(), "R"),
  destination_dir = file.path(getwd(), "dev"),
  filename = "combined.R",
  header_symbol = "=",
  max_new_lines = 0,
  new_lines = character(0)
)
}
\arguments{
\item{source_dir}{a file path for your source (such as R folder)}

\item{destination_dir}{a file path for your destination (such as dev folder)}

\item{filename}{a file name (ends in .R)}

\item{header_symbol}{single character for your header separator in combined file}

\item{max_new_lines}{integer for max number of new lines}

\item{new_lines}{characeter vector for new lines}
}
\value{
message
}
\description{
combine_R_files
}
