% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_diagram.R
\name{create_node_edge_REDCap}
\alias{create_node_edge_REDCap}
\title{create_node_edge_REDCap}
\usage{
create_node_edge_REDCap(DB, include_vars = F, type = "DiagrammeR")
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
messages for confirmation
}
\description{
create_node_edge_REDCap
}
