% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapSync-experimental.R
\name{add_project_transformation}
\alias{add_project_transformation}
\title{Add Default Forms Transformation to the Database}
\usage{
add_project_transformation(project, forms_transformation)
}
\arguments{
\item{forms_transformation}{a data.frame that matches instruments. See
`default_project_transformation` for an example.}
}
\value{
The updated `project` object with default transformations applied to the
specified forms.
}
\description{
`r lifecycle::badge("experimental")`
Applies default transformations to specific forms within the REDCap database
(`project`).
This function modifies the `project` object to include default
transformations, which may
involve adjustments, calculations, or reformatting of data in predefined
forms.
}
\details{
This function is designed to streamline and standardize data processing by
applying default transformations to the database forms. The transformations
are predefined within the function and ensure consistency across datasets.
}
\seealso{
\code{\link{save_project}} for saving the database or summaries.
}
