% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_all_functions.R
\name{deidentify_DB}
\alias{deidentify_DB}
\title{Deidentify the REDCap DB according to REDCap or your choices}
\usage{
deidentify_DB(DB, identifiers)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{identifiers}{optional character vector of column names that should be excluded from DB. Otherwise `DB$redcap$metadata$identifier =="y` will be used.}
}
\value{
DB object that has deidentified forms
}
\description{
Deidentify the REDCap DB according to REDCap or your choices
}
