% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_setup_DB.R
\name{delete_DB}
\alias{delete_DB}
\title{Deletes DB object from directory (solves occasional problems)}
\usage{
delete_DB(DB, dir_path)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{dir_path}{character file path of the directory}
}
\value{
message
}
\description{
Deletes DB object from directory (solves occasional problems)
}
\seealso{
Other DB object: 
\code{\link{load_DB}()},
\code{\link{save_DB}()},
\code{\link{setup_DB}()},
\code{\link{show_DB}()}
}
\concept{DB object}
