% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_files.R
\name{delete_file_from_redcap}
\alias{delete_file_from_redcap}
\title{delete_file_from_redcap}
\usage{
delete_file_from_redcap(
  DB,
  record,
  field,
  repeat_instance = NULL,
  event = NULL
)
}
\description{
delete_file_from_redcap
}
