% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyUtils.R
\name{DF_to_wb}
\alias{DF_to_wb}
\title{DF_to_wb}
\usage{
DF_to_wb(
  DF,
  DF_name,
  wb = openxlsx::createWorkbook(),
  link_col_list = list(),
  str_trunc_length = 32000,
  header_df,
  tableStyle = "none",
  header_style = default_header_style,
  body_style = default_body_style,
  freeze_header = T,
  pad_rows = 0,
  pad_cols = 0,
  freeze_keys = T,
  key_cols = NULL
)
}
\description{
DF_to_wb
}
