% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007_save_DB.R
\name{drop_DB_to_dir}
\alias{drop_DB_to_dir}
\title{drop_DB_to_dir}
\usage{
drop_DB_to_dir(
  DB,
  include_metadata = T,
  forms,
  str_trunc_length = 32000,
  separate = F,
  file_name
)
}
\value{
messages for confirmation
}
\description{
drop_DB_to_dir
}
