% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_process.R
\name{field_names_to_instruments}
\alias{field_names_to_instruments}
\title{field_names_to_instruments}
\usage{
field_names_to_instruments(DB, field_names, only_unique = T)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{only_unique}{logical for unique}
}
\value{
instrument names
}
\description{
field_names_to_instruments
}
