% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_process.R
\name{filter_DB}
\alias{filter_DB}
\title{Select REDCap records from DB}
\usage{
filter_DB(DB, records = NULL, data_choice = "data_extract", field_names)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{records}{character vector of the IDs you want to filter the DB by}
}
\value{
DB object that has been filtered to only include the specified records
}
\description{
Select REDCap records from DB
}
