% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_upload.R
\name{find_DB_diff}
\alias{find_DB_diff}
\title{Find the DB_import and DB differences}
\usage{
find_DB_diff(DB2, DB, ignore_instruments)
}
\arguments{
\item{DB2}{The comparison DB object}

\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{ignore_instruments}{character vector of instruments to be ignored if you know they are unchanged.}
}
\value{
DB_import but only the differences
}
\description{
This function is meant to be run after `DB_import <- read_redcap_dir(DB)`.
It compares DB_import to DB and only uploads the changes.
This will only overwrite and new data. It will not directly delete and data.
Because this is the only function that can mess up your data, use it at your own risk.
Remember all changes are saved in the redcap log if there's an issue.
}
