% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyUtils.R
\name{find_df_diff}
\alias{find_df_diff}
\title{find the difference between two data.frames}
\usage{
find_df_diff(new, old, ref_cols = NULL, message_pass = "")
}
\arguments{
\item{new}{a new data.frame to compare to old. All new cols must be included in the set of the old ones.}

\item{old}{a reference data.frame to be compared to}

\item{ref_cols}{character vector of reference columns. They are always included in the return data.frame and their combination should always lead to a unique key for each row.}
}
\value{
messages and data.frame of only changes and reference cols
}
\description{
This function will compare two data.frames: new and old.
You define the reference columns with ref_cols.
Reference columns are always included in the return data.frame and their combination should always lead to a unique key for each row.
}
