% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyUtils.R
\name{find_df_list_diff}
\alias{find_df_list_diff}
\title{find the difference between two lists of related data.frames}
\usage{
find_df_list_diff(
  new_list,
  old_list,
  ref_col_list,
  view_old = T,
  n_row_view = 20
)
}
\arguments{
\item{new_list}{a new data.frame to compare to old. All new cols must be included in the set of the old ones.}

\item{old_list}{a reference data.frame to be compared to}

\item{ref_col_list}{character vector of reference columns. They are always included in the return data.frame and their combination should always lead to a unique key for each row.}

\item{view_old}{logical for viewing old}
}
\value{
messages and data.frame of only changes and reference cols
}
\description{
This function will compare two data.frames: new and old.
You define the reference columns with ref_cols.
Reference columns are always included in the return data.frame and their combination should always lead to a unique key for each row.
}
