% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyUtils.R
\name{full.file.info}
\alias{full.file.info}
\title{full.file.info}
\usage{
full.file.info(path, showWarnings = T)
}
\arguments{
\item{path}{a file path}

\item{showWarnings}{logical for showing warnings}
}
\value{
file_info data.frame
}
\description{
full.file.info
}
