% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_transform.R
\name{generate_custom_remap_from_dir}
\alias{generate_custom_remap_from_dir}
\title{Generate custom remap files from input}
\usage{
generate_custom_remap_from_dir(DB)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
DB object that has DB$remap populated from input folder
}
\description{
Generate custom remap files from input
}
