% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_summarize_DB.R
\name{generate_summary_from_subset_name}
\alias{generate_summary_from_subset_name}
\title{generate_summary_from_subset_name}
\usage{
generate_summary_from_subset_name(
  DB,
  subset_name,
  clean = T,
  drop_blanks = T,
  other_drops = NULL,
  include_metadata = T,
  annotate_metadata = T,
  include_record_summary = T,
  include_users = T,
  include_log = T,
  add_to_global = T
)
}
\description{
generate_summary_from_subset_name
}
