% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_setup.R
\name{get_projects}
\alias{get_projects}
\title{Get your REDCap projects used by RosyREDCap}
\usage{
get_projects()
}
\value{
data.frame of projects from the cache
}
\description{
Everytime a setup or update is performed RosyREDCap stores the most basic information
about that project to the cache so the user has a running log of everywhere there project information is stored,
which can be used to find, move, edit, delete that data.
}
