% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_all_functions.R
\name{get_raw_redcap}
\alias{get_raw_redcap}
\title{Check the REDCap log}
\usage{
get_raw_redcap(DB, clean = T, records = NULL)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{clean}{T/F for clean vs raw labels}

\item{records}{optional records}
}
\value{
data.frame of raw_redcap
}
\description{
Check the REDCap log
}
