% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/109_REDCap_API.R
\name{get_REDCap_files}
\alias{get_REDCap_files}
\title{Drop redcap files to directory}
\usage{
get_REDCap_files(DB, original_file_names = F, overwrite = F)
}
\arguments{
\item{original_file_names}{logical for using original uploaded filenames vs system defined}

\item{overwrite}{logical rewriting over the downloaded files in your directory. A better alternative is deleting files you want to update.}
}
\value{
message
}
\description{
Drop redcap files to directory
}
