% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/109_REDCap_API.R
\name{get_REDCap_log}
\alias{get_REDCap_log}
\title{Check the REDCap Log}
\usage{
get_REDCap_log(
  DB,
  last = 24,
  user = "",
  units = "hours",
  begin_time = "",
  clean = T,
  record = ""
)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using `load_DB()` or `setup_DB()`}

\item{last}{Numeric. Specifies the time range paired with `units` to filter the log. Default is `24`.}

\item{user}{Character. Optional user filter to retrieve log entries associated with a specific username. Default is an empty string (`""`), which includes all users.}

\item{units}{Character. Units paired with `last` to define the time range. Options are `"mins"`, `"hours"`, `"days"`. Default is `"hours"`.}

\item{begin_time}{Character. A specific timestamp (e.g., "2023-07-11 13:15:06") to define the start of the log retrieval. Overrides `last` and `units` if provided. Default is an empty string (`""`).}

\item{clean}{Logical. If `TRUE`, cleans the API data and adds extra summary columns to the log. Default is `TRUE`.}

\item{record}{Character. Optional record filter for retrieving log entries associated with a specific record. Default is an empty string (`""`).}
}
\value{
A `data.frame` containing the cleaned and summarized REDCap log.
}
\description{
Retrieve and analyze the REDCap log to monitor recent activity or filter for specific users and time ranges.
}
\details{
This function queries the REDCap project log to retrieve recent activity based on the specified time range, user, or starting timestamp.
It optionally cleans the raw API data and adds summary columns to enhance log readability and analysis.
}
\seealso{
\code{\link[RosyREDCap]{update_DB}} for updating the `DB` object, including the log.
}
