% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_redcapAPI.R
\name{get_redcap_structure}
\alias{get_redcap_structure}
\title{get_redcap_structure}
\usage{
get_redcap_structure(DB, parse_codes = F)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{parse_codes}{logical for parsing choice codes and missing codes into tabs}
}
\value{
redcap structure list
}
\description{
get_redcap_structure
}
