% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_all_functions.R
\name{grab_record_tables}
\alias{grab_record_tables}
\title{grab data table for an individual(s)}
\usage{
grab_record_tables(DB, records)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{records}{character vector of the IDs you want to filter the DB by}
}
\value{
list of data tables
}
\description{
grab data table for an individual(s)
}
