% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_summarize.R
\name{labelled_to_raw_form}
\alias{labelled_to_raw_form}
\title{Clean to Raw REDCap forms}
\usage{
labelled_to_raw_form(FORM, DB)
}
\arguments{
\item{FORM}{data.frame of labelled REDCap to be converted to raw REDCap (for uploads)}

\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}
}
\value{
DB object that has been filtered to only include the specified records
}
\description{
Clean to Raw REDCap forms
}
