% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_links_from_DB.R
\name{link_API_playground}
\alias{link_API_playground}
\title{Link view the API playground for your project (if you access)}
\usage{
link_API_playground(DB)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
messages for confirmation
}
\description{
Link view the API playground for your project (if you access)
}
