% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_link.R
\name{link_REDCap_home}
\alias{link_REDCap_home}
\title{Link view the REDCap project home page}
\usage{
link_REDCap_home(DB)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
opens browser link
}
\description{
Link view the REDCap project home page
}
