% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyUtils.R
\name{list_to_excel}
\alias{list_to_excel}
\title{list_to_wb}
\usage{
list_to_excel(
  list,
  dir,
  file_name = NULL,
  separate = FALSE,
  overwrite = TRUE,
  link_col_list = list(),
  str_trunc_length = 32000,
  header_df_list,
  tableStyle = "none",
  header_style = default_header_style,
  body_style = default_body_style,
  freeze_header = T,
  pad_rows = 0,
  pad_cols = 0,
  freeze_keys = T,
  key_cols_list = NULL
)
}
\description{
list_to_wb
}
