% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_all_functions.R
\name{load_DB}
\alias{load_DB}
\title{Reads DB from the directory}
\usage{
load_DB(dir_path, blank = F)
}
\arguments{
\item{dir_path}{character file path of the directory}

\item{blank}{logical for blank load or last save}
}
\value{
DB
}
\description{
Reads DB from the directory
}
