% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_redcapAPI.R
\name{load_redcap_structure_from_dir}
\alias{load_redcap_structure_from_dir}
\title{load_redcap_structure_from_dir}
\usage{
load_redcap_structure_from_dir(
  DB,
  path = file.path(DB$dir_path, "REDCap", "REDCap_structure.xlsx")
)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{parse_codes}{logical for parsing choice codes and missing codes into tabs}
}
\value{
redcap structure list
}
\description{
load_redcap_structure_from_dir
}
