% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{make_parcats}
\alias{make_parcats}
\title{make_parcats}
\usage{
make_parcats(DF, remove_missing = FALSE, marginal_histograms = TRUE)
}
\arguments{
\item{DF}{data.frame}

\item{remove_missing}{Logical for removing missing/NA data}

\item{marginal_histograms}{Logical for including marginal_histograms}
}
\description{
make_parcats
}
