% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_make_table1.R
\name{make_table1}
\alias{make_table1}
\title{make_table1}
\usage{
make_table1(DB, df, group = "no_choice", variables, render.missing = F)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
messages for confirmation
}
\description{
make_table1
}
