% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_link.R
\name{nav_to_dir}
\alias{nav_to_dir}
\title{nav_to_dir}
\usage{
nav_to_dir(DB)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
opens browser link
}
\description{
nav_to_dir
}
