% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_bar}
\alias{plotly_bar}
\title{plotly_bar}
\usage{
plotly_bar(DF, x_col, y_col, name)
}
\arguments{
\item{DF}{data.frame}

\item{x_col}{column to be plotted for x}

\item{y_col}{column to be plotted for y}

\item{name}{title}
}
\value{
plotly
}
\description{
plotly_bar
}
