% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_parcats}
\alias{plotly_parcats}
\title{plotly_parcats}
\usage{
plotly_parcats(
  DF,
  remove_missing = TRUE,
  line_shape_curved = TRUE,
  numeric_to_factor = TRUE,
  quantiles = 5,
  more_descriptive_label = FALSE
)
}
\description{
plotly_parcats
}
