% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_projects.R
\name{project_health_check}
\alias{project_health_check}
\title{project_health_check}
\usage{
project_health_check()
}
\value{
project cache data.frame
}
\description{
Check directory, DB object, and REDCap token. Optional update.
}
\seealso{
Other Project Cache Functions: 
\code{\link{cache_clear}()},
\code{\link{cache_path}()},
\code{\link{get_projects}()}
}
\concept{Project Cache Functions}
\keyword{Cache}
\keyword{Functions}
\keyword{Project}
