% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007_save_DB.R
\name{read_DB_from_dir}
\alias{read_DB_from_dir}
\title{Reads DB from the dropped REDCap files in dir/REDCap/upload}
\usage{
read_DB_from_dir(
  DB,
  allow_all = T,
  drop_non_form_vars = T,
  stop_or_warn = "warn"
)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{allow_all}{logical TF for allowing DB$data names that are not also form names}

\item{drop_non_form_vars}{logical TF for dropping non-form variable names}

\item{stop_or_warn}{character string of whether to stop, warn, or do nothing when forbidden cols are present}

\item{drop_nonredcap_vars}{logical TF for dropping non-redcap variable names}
}
\value{
messages for confirmation
}
\description{
Reads DB from the dropped REDCap files in dir/REDCap/upload
}
