% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/112_to_and_from_dir.R
\name{read_from_REDCap_upload}
\alias{read_from_REDCap_upload}
\title{Reads DB from the dropped REDCap files in dir/REDCap/upload}
\usage{
read_from_REDCap_upload(
  DB,
  allow_all = T,
  drop_nonredcap_vars = T,
  drop_non_form_vars = T,
  stop_or_warn = "warn"
)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{allow_all}{logical TF for allowing DB$data names that are not also form names}

\item{drop_nonredcap_vars}{logical TF for dropping non-redcap variable names}

\item{drop_non_form_vars}{logical TF for dropping non-form variable names}

\item{stop_or_warn}{character string of whether to stop, warn, or do nothing when forbidden cols are present}
}
\value{
messages for confirmation
}
\description{
Reads DB from the dropped REDCap files in dir/REDCap/upload
}
