% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/109_REDCap_API.R
\name{get_REDCap_method}
\alias{get_REDCap_method}
\alias{get_REDCap_metadata}
\alias{get_REDCap_data}
\alias{get_REDCap_version}
\alias{get_REDCap_files}
\alias{get_REDCap_users}
\alias{get_REDCap_structure}
\alias{get_REDCap_report}
\alias{show_REDCap_API_methods}
\alias{show_REDCap_API_methods_table}
\title{REDCap API}
\usage{
get_REDCap_method(
  DB,
  method,
  error_action = "warn",
  additional_args = NULL,
  show_method_help = T
)

get_REDCap_metadata(DB, include_users = T)

get_REDCap_data(DB, labelled = T, records = NULL, batch_size = 2000)

get_REDCap_version(DB, show_method_help = T)

get_REDCap_files(DB, original_file_names = F, overwrite = F)

get_REDCap_users(DB)

get_REDCap_structure(DB)

get_REDCap_report(DB)

show_REDCap_API_methods()

show_REDCap_API_methods_table()
}
\arguments{
\item{original_file_names}{logical for using original uploaded filenames vs system defined}

\item{overwrite}{logical rewriting over the downloaded files in your directory. A better alternative is deleting files you want to update.}
}
\value{
DB object
}
\description{
REDCap API
}
