% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_diagram.R
\name{REDCap_diagramR}
\alias{REDCap_diagramR}
\title{REDCap_diagramR}
\usage{
REDCap_diagramR(DB, include_vars = F, render = T)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
messages for confirmation
}
\description{
REDCap_diagramR
}
