% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{rmarkdown_project}
\alias{rmarkdown_project}
\title{rmarkdown_project
`r lifecycle::badge("experimental")`}
\usage{
rmarkdown_project(project, dir_other)
}
\arguments{
\item{project}{project object from REDCapSync package}
}
\description{
Generate a rmarkdown PDF from a project object
}
\details{
You will need the [tinytex] and [reticulate] packages. Follow errors during your first use and afterwards it should work
}
