% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyREDCap-package.R
\docType{package}
\name{RosyREDCap-package}
\alias{RosyREDCap}
\alias{RosyREDCap-package}
\title{RosyREDCap: RosyREDCap}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Rosy development for your R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brandonerose/RosyREDCap}
  \item \url{https://brandonerose.github.io/RosyREDCap/}
  \item Report bugs at \url{https://github.com/brandonerose/RosyREDCap/issues}
}

}
\author{
\strong{Maintainer}: Brandon Rose \email{thecodingdocs@gmail.com} (\href{https://orcid.org/0009-0009-7813-1960}{ORCID})

Authors:
\itemize{
  \item Natalie Goulett
  \item Chris Hilsinger-Pate
  \item Nikola Susic
}

Other contributors:
\itemize{
  \item Sydney Stern (\href{https://orcid.org/0000-0002-1479-605X}{ORCID}) [contributor]
}

}
\keyword{internal}
