% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/116_summarize_DB.R
\name{run_quality_checks}
\alias{run_quality_checks}
\title{Run Quality Checks}
\usage{
run_quality_checks(DB)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \code{DB <- \link{load_DB}("PROJ")} or \link{setup_DB}()}
}
\description{
Run Quality Checks
}
