% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_setup_DB.R
\name{save_DB}
\alias{save_DB}
\alias{delete_DB}
\title{Save or Delete DB file from the directory}
\usage{
save_DB(DB)

delete_DB(DB)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}
}
\value{
Message
}
\description{
This will save/delete the "<short_name>_RosyREDCap.rdata" file in the given DB directories R_objects folder. These are optional functions given that `save_DB` is a also handled by a default parameter in `update_DB.`
}
\details{
delete_DB will not delete any other files from that directory. The user must delete any other files manually.
}
\seealso{
Other DB object: 
\code{\link{setup_DB}()}
}
\concept{DB object}
