% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_redcapAPI.R
\name{save_redcap_structure_to_dir}
\alias{save_redcap_structure_to_dir}
\title{save_redcap_structure_to_dir}
\usage{
save_redcap_structure_to_dir(DB, parse_codes = F)
}
\arguments{
\item{parse_codes}{logical for parsing choice codes and missing codes into tabs}
}
\value{
redcap structure list
}
\description{
save_redcap_structure_to_dir
}
