% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_token_management.R
\name{set_REDCap_token}
\alias{set_REDCap_token}
\title{Sets a valid token for this session}
\usage{
set_REDCap_token(DB)
}
\value{
message for confirmation
}
\description{
Sets a valid token for this session
}
