% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_all_functions.R
\name{show_DB}
\alias{show_DB}
\title{Shows DB in the env}
\usage{
show_DB(DB, also_metadata = T, transform = T)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{also_metadata}{logical for including metadata}

\item{transform}{show the data_transform section instead of data_extract}
}
\value{
DB tables
}
\description{
To add all of these to your global environment you could run the following code...
`data_list <- show_DB(DB)`
`for(NAME in names(data_list)){assign(NAME,data_list[[NAME]],pos = 1)}`
This code is not allowed on CRAN to avoid name conflicts
}
