% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_summarize_DB.R
\name{summarize_DB}
\alias{summarize_DB}
\title{Summarize DB}
\usage{
summarize_DB(
  DB,
  records = NULL,
  drop_blanks = T,
  data_choice = DB$internals$reference_state
)
}
\arguments{
\item{records}{character vector of records to be summarized}

\item{drop_blanks}{optional logical for dropping blanks}
}
\description{
Summarize DB
}
