% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_summarize_DB.R
\name{summarize_RosyREDCap}
\alias{summarize_RosyREDCap}
\title{summarize_RosyREDCap}
\usage{
summarize_RosyREDCap(
  DB,
  with_links = T,
  deidentify = T,
  clean = T,
  drop_blanks = T,
  drop_unknowns = T,
  include_metadata = T,
  annotate_metadata = T,
  include_record_summary = T,
  include_users = T,
  include_log = T,
  separate = F,
  force = F
)
}
\arguments{
\item{drop_blanks}{optional logical for dropping blanks}
}
\description{
summarize_RosyREDCap
}
