% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyUtils.R
\name{sync_dir}
\alias{sync_dir}
\title{sync_dir}
\usage{
sync_dir(from, to, top_level = T)
}
\arguments{
\item{from}{a file path for from}

\item{to}{a file path for to}

\item{top_level}{logical for being at top level of recursive function}
}
\value{
message
}
\description{
sync_dir
}
