% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_token_management.R
\name{test_REDCap_token_logical}
\alias{test_REDCap_token_logical}
\title{Test REDCap API token of DB object}
\usage{
test_REDCap_token_logical(DB)
}
\arguments{
\item{DB}{a list object that contains all metadata, data, log, transformations, and more for one REDCap project.}
}
\value{
TRUE or FALSE based on API token test
}
\description{
Test REDCap API token of DB object
}
