% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_transform.R
\name{transform_DB}
\alias{transform_DB}
\title{Transform DB}
\usage{
transform_DB(
  DB,
  merge_non_rep_to_reps = F,
  records = NULL,
  force = F,
  terminal_transformation = F
)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
DB object that has DB$data_transform, can be based on a remap file from input folder or default
}
\description{
Transform DB
}
