% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_update_DB.R
\name{update_DB}
\alias{update_DB}
\title{Shows DB in the env}
\usage{
update_DB(DB, force = F, day_of_log = 10, clean = T)
}
\arguments{
\item{DB}{DB from load_DB or setup_DB}

\item{force}{logical for force a fresh update}

\item{day_of_log}{numbers of days to be checked in the log}

\item{clean}{logical for whether or not to return raw or clean REDCap. Default is TRUE.}
}
\value{
messages for confirmation
}
\description{
Shows DB in the env
}
