% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_update.R
\name{update_RosyREDCap}
\alias{update_RosyREDCap}
\title{Shows DB in the env}
\usage{
update_RosyREDCap(
  DB,
  force = F,
  day_of_log = 10,
  labelled = T,
  get_files = F,
  original_file_names = F,
  entire_log = F,
  ask_about_overwrites = T
)
}
\arguments{
\item{DB}{DB from load_RosyREDCap or setup_DB}

\item{force}{logical for force a fresh update}

\item{day_of_log}{numbers of days to be checked in the log}

\item{labelled}{logical for whether or not to return raw or labelled REDCap. Default is TRUE.}

\item{get_files}{logical for whether or not to get files from redcap.}

\item{original_file_names}{logical for whether or not to use original file names.}
}
\value{
messages for confirmation
}
\description{
Shows DB in the env
}
