% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_redcap.R
\name{upload_DB_import_to_redcap}
\alias{upload_DB_import_to_redcap}
\title{Upload from your directory to REDCap}
\usage{
upload_DB_import_to_redcap(
  DB_import,
  DB,
  batch_size = 500,
  use_missing_codes = T
)
}
\arguments{
\item{DB_import}{obtained from your directory 'REDCap/upload' folder using, `DB_import <- read_redcap_dir(DB)`}

\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{batch_size}{numeric of how big the REDCap batch upload is. Default 500.}

\item{use_missing_codes}{logical for whether or not to check for missing codes. Default is TRUE.}
}
\value{
messages
}
\description{
This function is meant to be run after `DB_import <- read_redcap_dir(DB)`.
It compares DB_import to DB and only uploads the changes.
This will only overwrite and new data. It will not directly delete and data.
Because this is the only function that can mess up your data, use it at your own risk.
Remember all changes are saved in the redcap log if there's an issue.
}
