% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_files.R
\name{upload_file_to_REDCap}
\alias{upload_file_to_REDCap}
\title{upload_file_to_REDCap}
\usage{
upload_file_to_REDCap(
  DB,
  file,
  record,
  field,
  repeat_instance = NULL,
  event = NULL
)
}
\description{
upload_file_to_REDCap
}
