% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_file_repository.R
\name{upload_file_to_redcap_fileRepository}
\alias{upload_file_to_redcap_fileRepository}
\title{Uploads a file to REDCap}
\usage{
upload_file_to_redcap_fileRepository(DB, file)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{file}{file location on your PC}
}
\value{
messages for confirmation
}
\description{
Uploads a file to REDCap
}
