% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/113_upload.R
\name{upload_form_to_REDCap}
\alias{upload_form_to_REDCap}
\title{Upload to REDCap}
\usage{
upload_form_to_REDCap(to_be_uploaded, DB, batch_size = 500)
}
\arguments{
\item{to_be_uploaded}{data.frame in raw coded form. If you worked with clean data pass your data to `labelled_to_raw_form(FORM,DB)` first.}

\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{batch_size}{numeric of how big the REDCap batch upload is. Default 500.}
}
\value{
messages
}
\description{
This will only overwrite and new data. It will not directly delete and data.
Because this is the only function that can mess up your data, use it at your own risk.
Remember all changes are saved in the redcap log if there's an issue. Missing rows and columns are fine!
}
