% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_redcapAPI.R
\name{upload_redcap_structure}
\alias{upload_redcap_structure}
\title{upload_redcap_structure}
\usage{
upload_redcap_structure(DB, redcap)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{redcap}{list to be uploaded}
}
\value{
redcap structure list
}
\description{
upload_redcap_structure
}
