% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_transform.R
\name{upload_transform_to_DB}
\alias{upload_transform_to_DB}
\title{upload_transform_to_DB Transform}
\usage{
upload_transform_to_DB(DB)
}
\arguments{
\item{DB}{a list object that contains all metadata, data, log, transformations, and more for one REDCap project.}
}
\description{
upload_transform_to_DB Transform
}
