% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyUtils.R
\name{warn_or_stop}
\alias{warn_or_stop}
\title{warn_or_stop}
\usage{
warn_or_stop(m, warn_only = F)
}
\arguments{
\item{m}{message character string}

\item{warn_only}{logical for only warn}
}
\value{
message
}
\description{
warn_or_stop
}
