% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_file_repository.R
\name{add_redcap_folder}
\alias{add_redcap_folder}
\title{Uploads a folder name to REDCap}
\usage{
add_redcap_folder(DB, name)
}
\arguments{
\item{name}{folder name}
}
\value{
messages for confirmation
}
\description{
Uploads a folder name to REDCap
}
