% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110_files.R
\name{add_redcap_folder}
\alias{add_redcap_folder}
\title{Uploads a folder name to REDCap}
\usage{
add_redcap_folder(DB, name)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \code{DB <- \link{load_DB}("PROJ")} or \link{setup_DB}()}

\item{name}{folder name}
}
\value{
messages for confirmation
}
\description{
Uploads a folder name to REDCap
}
