% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110_files.R
\name{add_REDCap_folder}
\alias{add_REDCap_folder}
\title{Uploads a folder name to REDCap}
\usage{
add_REDCap_folder(DB, name)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{name}{folder name}
}
\value{
messages for confirmation
}
\description{
Uploads a folder name to REDCap
}
