% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_setup.R
\name{cache_clear}
\alias{cache_clear}
\title{Clear your cached projects}
\usage{
cache_clear()
}
\value{
The file path of your RosyREDCap cache

your path
}
\description{
Included for transparency and confirmation/testing.
}
\details{
This function checks the location of the cache established by [hoardr::hoard()] and deletes it!
This will not delete project data, just the packages "memory" of it.
}
\examples{
\dontrun{
#warning this will delete the packages cache of project locations
cache_clear()
}
}
\seealso{
Other Cache Functions: 
\code{\link{cache_path}()}
}
\concept{Cache Functions}
\keyword{cache}
