% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_setup.R
\name{cache_clear}
\alias{cache_clear}
\title{Clear your cached projects}
\usage{
cache_clear()
}
\value{
message confirmation
}
\description{
Clear your cached projects
}
\seealso{
Other Cache: 
\code{\link{cache_path}()}
}
\concept{Cache}
