% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_directories.R
\name{check_folder_for_projects}
\alias{check_folder_for_projects}
\title{list file paths of any RosyREDCap in a folder}
\usage{
check_folder_for_projects(file_path, validate = T)
}
\arguments{
\item{validate}{logical T/F. If T function will only accept valid directories previously setup with `setup_DB()`}
}
\value{
character vector file paths
}
\description{
list file paths of any RosyREDCap in a folder
}
