% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_file_repository.R
\name{check_REDCap_files}
\alias{check_REDCap_files}
\title{Checks REDCap for current files}
\usage{
check_REDCap_files(DB)
}
\value{
data.frame of files
}
\description{
Checks REDCap for current files
}
