% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_file_repository.R
\name{check_redcap_files}
\alias{check_redcap_files}
\title{Checks REDCap for current files}
\usage{
check_redcap_files(DB)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
data.frame of files
}
\description{
Checks REDCap for current files
}
