% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_redcapAPI.R
\name{check_REDCap_log}
\alias{check_REDCap_log}
\title{Check the REDCap log}
\usage{
check_REDCap_log(
  DB,
  last = 24,
  user = "",
  units = "hours",
  begin_time = "",
  clean = T,
  record = ""
)
}
\arguments{
\item{last}{numeric paired with units. Default is 24.}

\item{user}{optional user filter.}

\item{units}{character paired with last. Options are "mins","hours","days". Default is "hours".}

\item{begin_time}{character of time where the log should start from. Example 2023-07-11 13:15:06.}

\item{clean}{logical for cleaning of API data}
}
\value{
data.frame of log that has been cleaned and has extra summary columns
}
\description{
Check the REDCap log
}
