% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_annotate_DB.R
\name{clean_DB}
\alias{clean_DB}
\title{clean DB columns for plotting using the metadata}
\usage{
clean_DB(DB, drop_blanks = F, other_drops = NULL)
}
\arguments{
\item{drop_blanks}{logical for dropping n=0 choices}

\item{drop_unknowns}{logical for dropping missing codes}
}
\value{
DB object cleaned for table or plots
}
\description{
Turns choices into factors and integers to integer for table processing such as with table1 and plots
}
