% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_setup_DB.R
\name{delete_DB}
\alias{delete_DB}
\title{Deletes DB object from directory (solves occasional problems)}
\usage{
delete_DB(DB, dir_path)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{dir_path}{character file path of the directory}
}
\value{
message
}
\description{
This will delete the "PROJ_RosyREDCap.rdata" file in the given DB directories R_objects folder. It will NOT delete any other files from that directory. The user must delete any other files manually.
}
\seealso{
Other DB object: 
\code{\link{save_DB}()},
\code{\link{setup_DB}()}
}
\concept{DB object}
