% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110_files.R
\name{delete_file_from_REDCap}
\alias{delete_file_from_REDCap}
\title{delete_file_from_REDCap}
\usage{
delete_file_from_REDCap(
  DB,
  record,
  field,
  repeat_instance = NULL,
  event = NULL
)
}
\description{
delete_file_from_REDCap
}
