% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110_files.R
\name{delete_REDCap_file}
\alias{delete_REDCap_file}
\title{Uploads a folder name to REDCap}
\usage{
delete_REDCap_file(DB, doc_id)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{doc_id}{from the file list `get_REDCap_file_repository(DB)`}
}
\value{
messages for confirmation
}
\description{
Uploads a folder name to REDCap
}
