% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_file_repository.R
\name{delete_redcap_file}
\alias{delete_redcap_file}
\title{Uploads a folder name to REDCap}
\usage{
delete_redcap_file(DB, doc_id)
}
\arguments{
\item{doc_id}{from the file list `check_REDCap_files(DB)`}
}
\value{
messages for confirmation
}
\description{
Uploads a folder name to REDCap
}
