% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_to_and_from_dir.R
\name{drop_redcap_dir}
\alias{drop_redcap_dir}
\title{Shows DB in the env}
\usage{
drop_redcap_dir(
  DB,
  smart = T,
  include_metadata = T,
  include_other = T,
  with_links = F,
  forms,
  merge_non_repeating = T,
  separate = F
)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{smart}{logical for whether to only save when data is new}

\item{include_metadata}{logical for whether to only include redcap and not metadata}

\item{include_other}{logical for whether to only include redcap and not metadata}

\item{with_links}{optional logical for including links in excel sheets}

\item{forms}{optional character vector for only selected forms}

\item{merge_non_repeating}{optional logical for merging non-repeating instruments}

\item{records}{character vector of records you want dropped to your directory}

\item{allow_mod}{logical for whether non-instrument names are allowed}

\item{deidentify}{logical for deidentification}

\item{dir_other}{optional character string of another file path where the files should be saved}

\item{append_name}{optional character string for adding to the front of file names}

\item{str_trunc_length}{optional integer for truncation}
}
\value{
messages for confirmation
}
\description{
Shows DB in the env
}
