% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_redcap_to_and_from_dir.R
\name{drop_redcap_dir}
\alias{drop_redcap_dir}
\title{Shows DB in the env}
\usage{
drop_redcap_dir(DB, records = NULL, allow_mod = T)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{records}{character vector of records you want dropped to your directory}

\item{allow_mod}{logical for whether non-instrument names are allowed}
}
\value{
messages for confirmation
}
\description{
Shows DB in the env
}
