% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_to_and_from_dir.R
\name{drop_redcap_dir}
\alias{drop_redcap_dir}
\title{Shows DB in the env}
\usage{
drop_redcap_dir(
  DB,
  records,
  allow_mod = T,
  dir_other,
  smart = T,
  include_metadata = T,
  include_other = F,
  deidentify = F,
  append_name,
  str_trunc_length = 32000,
  with_links = T,
  annotate_codebook = T,
  forms
)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{records}{character vector of records you want dropped to your directory}

\item{allow_mod}{logical for whether non-instrument names are allowed}

\item{dir_other}{optional character string of another file path where the files should be saved}

\item{smart}{logical for whether to only save when data is new}

\item{include_metadata}{logical for whether to only include redcap and not metadata}

\item{include_other}{logical for whether to only include redcap and not metadata}

\item{deidentify}{logical for deidentification}

\item{append_name}{optional character string for adding to the front of file names}

\item{str_trunc_length}{optional integer for truncation}

\item{with_links}{optional logical for including links in excel sheets}

\item{annotate_codebook}{optional logical for adding annotations of n and percent to codebook file}

\item{forms}{optional character vector for only selected forms}
}
\value{
messages for confirmation
}
\description{
Shows DB in the env
}
