% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_redcap.R
\name{drop_redcap_dir}
\alias{drop_redcap_dir}
\title{Shows DB in the env}
\usage{
drop_redcap_dir(DB)
}
\arguments{
\item{DB}{list for the package that contains applications and reference files}
}
\value{
messages for confirmation
}
\description{
Shows DB in the env
}
