% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/112_to_and_from_dir.R
\name{drop_REDCap_to_directory}
\alias{drop_REDCap_to_directory}
\title{Drop REDCap Data to Directory}
\usage{
drop_REDCap_to_directory(
  DB,
  records,
  smart = T,
  include_metadata = T,
  include_other = T,
  with_links = T,
  forms,
  merge_non_repeating = T,
  separate = F,
  str_trunc_length = 32000,
  file_name
)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \code{DB <- \link{load_DB}("PROJ")} or \link{setup_DB}()}

\item{records}{Character vector of records you want dropped to your directory.}

\item{smart}{Logical (TRUE/FALSE). If TRUE, only saves when data is new. Default is `TRUE`.}

\item{include_metadata}{Logical (TRUE/FALSE). If TRUE, includes metadata in the export. Default is `TRUE`.}

\item{include_other}{Logical (TRUE/FALSE). If TRUE, includes other data in the export. Default is `TRUE`.}

\item{with_links}{Optional logical (TRUE/FALSE) for including links in Excel sheets. Default is `FALSE`.}

\item{forms}{Optional character vector for selecting specific forms to export.}

\item{merge_non_repeating}{Optional logical (TRUE/FALSE) for merging non-repeating instruments. Default is `FALSE`.}

\item{str_trunc_length}{Optional integer for truncation of strings.}

\item{file_name}{Optional character string for adding to the front of file names.}

\item{allow_mod}{Logical (TRUE/FALSE). If TRUE, allows non-instrument names. Default is `FALSE`.}

\item{deidentify}{Logical (TRUE/FALSE). If TRUE, deidentifies the data. Default is `FALSE`.}

\item{dir_other}{Optional character string of another file path where the files should be saved.}
}
\value{
Messages for confirmation.
}
\description{
Exports specified records from the REDCap database (`DB` object) to the specified directory.
}
\details{
This function exports the specified records from the REDCap database to the specified directory. It supports various options such as deidentification, including metadata, and merging non-repeating instruments. The function can also save the data only when it is new, and it allows for customization of file names and truncation of strings.
}
\seealso{
\code{\link[RosyREDCap]{setup_RosyREDCap}} for initializing the `DB` object.
}
\concept{export_functions}
