% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/113_upload.R
\name{edit_redcap_while_viewing}
\alias{edit_redcap_while_viewing}
\title{edit_redcap_while_viewing}
\usage{
edit_redcap_while_viewing(
  DB,
  optional_DF,
  records,
  field_name_to_change,
  field_names_to_view = NULL,
  upload_individually = T
)
}
\description{
edit_redcap_while_viewing
}
