% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_process.R
\name{filter_DB}
\alias{filter_DB}
\title{Select REDCap records from DB}
\usage{
filter_DB(
  DB,
  records,
  data_choice = "data",
  field_names,
  form_names,
  add_filter_var,
  add_filter_vals,
  warn_only = F
)
}
\arguments{
\item{records}{character vector of the IDs you want to filter the DB by}

\item{data_choice}{character vector of the IDs you want to filter the DB by}

\item{field_names}{character vector of field_names to be included}

\item{form_names}{character vector of form_names to be included}

\item{add_filter_var}{character string of extra variable name to be filtered by if present in a data.frame}

\item{add_filter_vals}{character vector of extra variable values to be filtered by if present in a data.frame}

\item{warn_only}{logical for warn_only or stop}
}
\value{
DB object that has been filtered to only include the specified records
}
\description{
Select REDCap records from DB
}
