% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006_summarize_DB.R
\name{filter_DB}
\alias{filter_DB}
\title{Select REDCap records from DB}
\usage{
filter_DB(
  DB,
  filter_field,
  filter_choices,
  form_names,
  field_names,
  warn_only = F,
  no_duplicate_cols = F
)
}
\arguments{
\item{filter_field}{character string of extra variable name to be filtered by if present in a data.frame}

\item{filter_choices}{character vector of extra variable values to be filtered by if present in a data.frame}

\item{form_names}{character vector of form_names to be included}

\item{field_names}{character vector of field_names to be included}

\item{warn_only}{logical for warn_only or stop}
}
\value{
DB object that has been filtered to only include the specified records
}
\description{
Select REDCap records from DB
}
