% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_upload.R
\name{find_upload_diff}
\alias{find_upload_diff}
\title{Find the DB_import and DB differences}
\usage{
find_upload_diff(DB, view_old = F, n_row_view = 20)
}
\arguments{
\item{compare}{what data_choice to be compare}

\item{to}{what data_choice to be compared to}
}
\value{
upload_list
}
\description{
Find the DB_import and DB differences
}
