% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/115_transform.R
\name{generate_horizontal_transform}
\alias{generate_horizontal_transform}
\title{Horizontal Transform}
\usage{
generate_horizontal_transform(DB, records)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \code{DB <- \link{load_DB}("PROJ")} or \link{setup_DB}()}
}
\description{
Horizontal Transform
}
