% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_transform.R
\name{generate_horizontal_transform}
\alias{generate_horizontal_transform}
\title{Horizontal Transform}
\usage{
generate_horizontal_transform(DB, records)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{records}{character vector of records to be summarized}
}
\description{
Horizontal Transform
}
