% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_transform.R
\name{generate_horizontal_transform}
\alias{generate_horizontal_transform}
\title{Horizontal Transform}
\usage{
generate_horizontal_transform(DB, records)
}
\arguments{
\item{DB}{a list object slightly modified version of RosyDB and contains all metadata, data, transformations, and more}
}
\description{
Horizontal Transform
}
