% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_transform.R
\name{generate_horizontal_transform}
\alias{generate_horizontal_transform}
\title{Horizontal Transform}
\usage{
generate_horizontal_transform(DB, records)
}
\arguments{
\item{DB}{a list object that contains all metadata, data, log, transformations, and more for one REDCap project.}
}
\description{
Horizontal Transform
}
