% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_directories.R
\name{get_dir}
\alias{get_dir}
\title{get your directory}
\usage{
get_dir(DB)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}
}
\value{
file path of directory
}
\description{
get your directory
}
