% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/109_REDCap_API.R
\name{get_raw_REDCap}
\alias{get_raw_REDCap}
\title{Check the REDCap log}
\usage{
get_raw_REDCap(DB, labelled = F, records = NULL, batch_size = 1000)
}
\arguments{
\item{labelled}{T/F for clean vs raw labels}

\item{records}{optional records}
}
\value{
data.frame of raw_redcap
}
\description{
Check the REDCap log
}
