% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110_files.R
\name{get_REDCap_file_repository}
\alias{get_REDCap_file_repository}
\title{Checks REDCap for current files}
\usage{
get_REDCap_file_repository(DB)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \code{DB <- \link{load_DB}("PROJ")} or \link{setup_DB}()}
}
\value{
data.frame of files
}
\description{
Checks REDCap for current files
}
