% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110_files.R
\name{get_REDCap_file_repository}
\alias{get_REDCap_file_repository}
\title{Checks REDCap for current files}
\usage{
get_REDCap_file_repository(DB)
}
\value{
data.frame of files
}
\description{
Checks REDCap for current files
}
