% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_redcapAPI.R
\name{get_redcap_files}
\alias{get_redcap_files}
\title{Drop redcap files to directory}
\usage{
get_redcap_files(DB, original_file_names = F, overwrite = F)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{original_file_names}{logical for using original uploaded filenames vs system defined}

\item{overwrite}{logical rewriting over the downladed files in your directory. A better alternative is deleting files you want to update.}
}
\value{
message
}
\description{
Drop redcap files to directory
}
