% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/109_REDCap_API.R
\name{get_REDCap_raw_data}
\alias{get_REDCap_raw_data}
\title{Check the REDCap log}
\usage{
get_REDCap_raw_data(DB, labelled = F, records = NULL, batch_size = 1000)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{labelled}{Logical. If `TRUE`, the function retrieves labeled data; if `FALSE`, it retrieves raw data. Default is `FALSE`.}

\item{records}{Optional character vector of record IDs to retrieve. If `NULL`, all records are fetched.}

\item{batch_size}{Numeric. The batch size for the retrieval process. Default is `1000`.}
}
\value{
A `data.frame` containing the raw or labeled REDCap data.
}
\description{
Retrieves raw or labeled data from a REDCap database based on the provided parameters.
The function connects to a REDCap project using the REDCap API, fetches the raw or labeled data
(depending on the `labelled` parameter), and returns the data as a `data.frame`. Optionally,
specific records can be retrieved.
}
\details{
This function uses the \code{\link[REDCapR]{redcap_read}} function to fetch data from a REDCap project.
The function allows for batch retrieval, and the `labelled` argument determines whether the
data is returned with raw values or with labels. By default, the function retrieves raw data.
It also supports filtering for specific records if provided.
}
\seealso{
\code{\link[REDCapR]{redcap_read}} for details on the REDCap API function used to fetch data.

Other db_functions: 
\code{\link{rmarkdown_DB}()},
\code{\link{summarize_DB}()},
\code{\link{transform_DB}()},
\code{\link{update_DB}()}
}
\concept{db_functions}
