% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_redcapAPI.R
\name{get_REDCap_structure}
\alias{get_REDCap_structure}
\title{get_REDCap_structure}
\usage{
get_REDCap_structure(DB, parse_codes = F)
}
\arguments{
\item{parse_codes}{logical for parsing choice codes and missing codes into tabs}
}
\value{
redcap structure list
}
\description{
get_REDCap_structure
}
