% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_process.R
\name{labelled_to_raw_form}
\alias{labelled_to_raw_form}
\title{Clean to Raw REDCap forms}
\usage{
labelled_to_raw_form(FORM, DB)
}
\arguments{
\item{FORM}{data.frame of labelled REDCap to be converted to raw REDCap (for uploads)}

\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
DB object that has been filtered to only include the specified records
}
\description{
Clean to Raw REDCap forms
}
