% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/117_link.R
\name{link_API_token}
\alias{link_API_token}
\title{Link to get a new API token for your project (if you access)}
\usage{
link_API_token(DB)
}
\value{
messages for confirmation
}
\description{
Link to get a new API token for your project (if you access)
}
