% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_link.R
\name{link_REDCap_project}
\alias{link_REDCap_project}
\title{link_REDCap_project}
\usage{
link_REDCap_project(DB)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
opens browser link
}
\description{
link_REDCap_project
}
