% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/117_link.R
\name{link_REDCap_record}
\alias{link_REDCap_record}
\title{link_REDCap_record}
\usage{
link_REDCap_record(DB, record, page, instance, text_only = F)
}
\arguments{
\item{record}{REDCap record id or study id etc, any column names that match `DB$redcap$id_col`}

\item{page}{REDCap page for the record. Must be one of `DB$metadata$forms$form_name`}

\item{instance}{REDCap instance if it's a repeating instrument}

\item{text_only}{logical for only returning text}
}
\value{
opens browser link
}
\description{
link_REDCap_record
}
