% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_link.R
\name{link_REDCap_record}
\alias{link_REDCap_record}
\title{Shows DB in the env}
\usage{
link_REDCap_record(DB, record, page, instance)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{record}{REDCap record id or study id etc, any column names that match `DB$redcap$id_col`}

\item{page}{REDCap page for the record. Must be one of `DB$redcap$instruments$instrument_name`}

\item{instance}{REDCap instance if it's a repeating instrument}
}
\value{
opens browser link
}
\description{
Shows DB in the env
}
