% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_setup_DB.R
\name{load_DB}
\alias{load_DB}
\title{Load RosyREDCap Project}
\usage{
load_DB(short_name, dir_path, validate = T)
}
\arguments{
\item{short_name}{A character string representing the short name you previously chose for the REDCap project.}

\item{dir_path}{A character string representing the directory path where the REDCap project files are stored. Optional if already set by `setup_DB()`.}

\item{validate}{Logical (TRUE/FALSE). If TRUE, validates the loaded `DB` object. Default is `TRUE`.}
}
\value{
The RosyREDCap `DB` list object.
}
\description{
Will take your previously chosen `short_name` and load the directory-saved DB object by using the cache. `dir_path` is optional if you already used `setup_DB()`.
}
\details{
This function loads the `DB` object from the specified directory using the provided `short_name`. If `dir_path` is not provided, it will use the path set during the `setup_DB()` call.
Optionally, it can validate the loaded `DB` object. It references a cache that is updated whenever you run `setup_DB()`.
}
\examples{

\dontrun{
# Load the DB object using the short name and directory path
DB <- load_DB("ABC")
}
}
\seealso{
\code{\link[RosyREDCap]{get_projects}} for retrieving a list of projects from the directory cache.
\code{\link[RosyREDCap]{setup_DB}} for retrieving a list of projects from the directory cache.

Other DB object: 
\code{\link{delete_DB}()},
\code{\link{save_DB}()},
\code{\link{setup_DB}()},
\code{\link{show_DB}()}
}
\concept{DB object}
