% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_db.R
\name{load_DB}
\alias{load_DB}
\title{Reads DB from the directory}
\usage{
load_DB(blank = F)
}
\arguments{
\item{blank}{logical for blank load or last save}
}
\value{
DB
}
\description{
Reads DB from the directory
}
