% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_DB_mod.R
\name{load_RosyREDCap}
\alias{load_RosyREDCap}
\title{Load RosyREDCap Project}
\usage{
load_RosyREDCap(short_name, dir_path, validate = T)
}
\arguments{
\item{short_name}{A character string representing the short name you previously chose for the REDCap project.}

\item{dir_path}{A character string representing the directory path where the REDCap project files are stored. Optional if already set by `setup_RosyREDCap()`.}

\item{validate}{Logical (TRUE/FALSE). If TRUE, validates the loaded `DB` object. Default is `TRUE`.}
}
\value{
The RosyREDCap `DB` list object.
}
\description{
Will take your previously chosen `short_name` and load the directory-saved DB object by using the cache. `dir_path` is optional if you already used `setup_RosyREDCap()`.
}
\details{
This function loads the `DB` object from the specified directory using the provided `short_name`. If `dir_path` is not provided, it will use the path set during the `setup_RosyREDCap()` call.
Optionally, it can validate the loaded `DB` object. It references a cache that is updated whenever you run `setup_RosyREDCap()`.
}
\examples{
\dontrun{
# Load the DB object using the short name and directory path
DB <- load_RosyREDCap("ABC")
}
}
\seealso{
\code{\link[RosyDB]{get_projects}} for retrieving a list of projects from the directory cache.
\code{\link[RosyREDCap]{setup_RosyREDCap}} for retrieving a list of projects from the directory cache.
}
