% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{make_survival}
\alias{make_survival}
\title{make_survival}
\usage{
make_survival(
  DF,
  start_col,
  end_col,
  units = "months",
  time_col,
  status_col,
  strat_col,
  allowed_levels,
  pval = TRUE,
  pval.coord = c(0, 0.25),
  risk.table = TRUE,
  title = NULL,
  palette = NULL,
  conf.int = TRUE,
  xlim = NULL,
  legend.position = "top",
  tables.height = 0.25,
  show_stats = FALSE
)
}
\description{
make_survival
}
