% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_table1}
\alias{make_table1}
\title{make_table1}
\usage{
make_table1(DF, group = "no_choice", variables = NULL, render.missing = FALSE)
}
\arguments{
\item{DF}{data.frame}

\item{group}{character string of no_choice or a factor column name}

\item{variables}{character vector of variables to include. NULL is all.}

\item{render.missing}{logical for rendering missing data or not}
}
\description{
make_table1
}
