% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/002_directories.R
\name{nav_to_dir}
\alias{nav_to_dir}
\title{nav_to_dir}
\usage{
nav_to_dir(DB)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}
}
\value{
opens browser link
}
\description{
nav_to_dir
}
