% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_parcats}
\alias{plotly_parcats}
\title{plotly_parcats}
\usage{
plotly_parcats(
  DF,
  remove_missing = TRUE,
  line_shape_curved = TRUE,
  numeric_to_factor = TRUE,
  quantiles = 5L,
  more_descriptive_label = FALSE
)
}
\arguments{
\item{DF}{data.frame}

\item{remove_missing}{logical for removing missing/NA data}

\item{line_shape_curved}{logical for having curved lines}

\item{numeric_to_factor}{logical for making numerics factors using quantiles}

\item{quantiles}{integer for quantiles}

\item{more_descriptive_label}{logical for having more descriptive labes}
}
\value{
plotly
}
\description{
plotly_parcats
}
