% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plotly_parcats}
\alias{plotly_parcats}
\title{plotly_parcats}
\usage{
plotly_parcats(
  DF,
  remove_missing = T,
  line_shape_curved = T,
  numeric_to_factor = T,
  quantiles = 5,
  more_descriptive_label = F
)
}
\description{
plotly_parcats
}
