% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/006_summarize_DB.R
\name{raw_to_labelled_form}
\alias{raw_to_labelled_form}
\title{Raw to Labelled REDCap forms}
\usage{
raw_to_labelled_form(FORM, DB)
}
\arguments{
\item{FORM}{data.frame of raw REDCap to be converted to labelled REDCap}
}
\value{
DB object
}
\description{
Raw to Labelled REDCap forms
}
