% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_redcap _to_and_from_dir.R
\name{read_redcap_dir}
\alias{read_redcap_dir}
\title{Reads DB from the dropped REDCap files in dir/REDCap/upload}
\usage{
read_redcap_dir(DB)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
messages for confirmation
}
\description{
Reads DB from the dropped REDCap files in dir/REDCap/upload
}
