% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_to_and_from_dir.R
\name{read_redcap_dir}
\alias{read_redcap_dir}
\title{Reads DB from the dropped REDCap files in dir/REDCap/upload}
\usage{
read_redcap_dir(
  DB,
  allow_all = T,
  drop_nonredcap_vars = T,
  drop_non_instrument_vars = T,
  stop_or_warn = "warn"
)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}

\item{allow_all}{logical TF for allowing DB$data_extract names that are not also instrument names}

\item{drop_nonredcap_vars}{logical TF for dropping non-redcap variable names}

\item{drop_non_instrument_vars}{logical TF for dropping non-instrument variable names}

\item{stop_or_warn}{character string of whether to stop, warn, or do nothing when forbidden cols are present}
}
\value{
messages for confirmation
}
\description{
Reads DB from the dropped REDCap files in dir/REDCap/upload
}
