% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_diagram.R
\name{REDCap_diagram}
\alias{REDCap_diagram}
\title{REDCap_diagram}
\usage{
REDCap_diagram(
  DB,
  render = T,
  include_fields = F,
  type = "visNetwork",
  duplicate_forms = T,
  clean_name = T
)
}
\value{
messages for confirmation
}
\description{
REDCap_diagram
}
