% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_diagram.R
\name{REDCap_diagram}
\alias{REDCap_diagram}
\title{REDCap_diagram}
\usage{
REDCap_diagram(DB, include_vars = F, type = "visNetwork", render = T)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
messages for confirmation
}
\description{
REDCap_diagram
}
