% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/118_diagram.R
\name{REDCap_diagram}
\alias{REDCap_diagram}
\title{REDCap_diagram}
\usage{
REDCap_diagram(
  DB,
  static = F,
  render = T,
  duplicate_forms = T,
  clean_names = T,
  include_fields = F,
  include_choices = F,
  hierarchical = F,
  direction = "LR"
)
}
\value{
messages for confirmation
}
\description{
REDCap_diagram
}
