% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{rmarkdown_project}
\alias{rmarkdown_project}
\title{rmarkdown_project}
\usage{
rmarkdown_project(project, dir_other)
}
\arguments{
\item{project}{project object from REDCapSync package}

\item{dir_other}{Character. The directory where the dataset file will be
saved. Default is the \code{output} folder within the database directory.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Generate a rmarkdown PDF from a project object
}
\details{
You will need the \code{tinytex} and \code{reticulate} packages. Follow errors during
your first use and afterwards it should work. Currently working on mac but
troubleshooting windows errors with latex.
}
