% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapSync-experimental.R
\name{rmarkdown_project}
\alias{rmarkdown_project}
\title{rmarkdown_project}
\usage{
rmarkdown_project(project, dir_other)
}
\arguments{
\item{dir_other}{Character string specifying the directory where the
RMarkdown report will be saved. If not provided, it defaults to the `output`
directory inside the project's main directory.}
}
\value{
A message indicating the creation of the RMarkdown report and the
path to the generated file.
}
\description{
`r lifecycle::badge("experimental")`
Generates an RMarkdown report for the given REDCap database (`project`
object). This function creates an RMarkdown file in the specified directory
or default directory, allowing users to create custom reports based on the
database content.
}
\details{
This function checks if a directory is specified, and if not, defaults to the
`output` folder within the project's directory. It generates the RMarkdown
file that can then be used for further processing or rendering into HTML,
PDF, or other formats.
}
\seealso{
\code{\link[REDCapSync]{save_project}} for saving the `project` object.
}
\concept{db_functions}
