% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RosyREDCap-package.R
\docType{package}
\name{RosyREDCap-package}
\alias{RosyREDCap}
\alias{RosyREDCap-package}
\title{RosyREDCap: Standardized REDCap Shiny Application}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Uses REDCapSync to load or build standardized R list objects for data viewing, quaility, and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thecodingdocs/RosyREDCap}
  \item \url{https://thecodingdocs.github.io/RosyREDCap/}
  \item Report bugs at \url{https://github.com/thecodingdocs/RosyREDCap/issues}
}

}
\author{
\strong{Maintainer}: Brandon Rose \email{thecodingdocs@gmail.com} (\href{https://orcid.org/0009-0009-7813-1960}{ORCID})

Other contributors:
\itemize{
  \item Natalie Goulett [contributor]
  \item Chris Hilsinger-Pate [contributor]
  \item Nikola Susic [contributor]
  \item Sydney Stern (\href{https://orcid.org/0000-0002-1479-605X}{ORCID}) [contributor]
}

}
\keyword{internal}
