% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/116_summarize_DB.R
\name{run_quality_checks}
\alias{run_quality_checks}
\title{Run Quality Checks}
\usage{
run_quality_checks(DB)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\description{
Run Quality Checks
}
