% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_summarize.R
\name{run_quality_checks}
\alias{run_quality_checks}
\title{Run Quality Checks}
\usage{
run_quality_checks(DB)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}
}
\value{
DB object
}
\description{
Run Quality Checks
}
