% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_setup_DB.R
\name{save_DB}
\alias{save_DB}
\title{Saves DB in the directory}
\usage{
save_DB(DB)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}
}
\value{
Message
}
\description{
Saves DB in the directory
}
\seealso{
Other DB object: 
\code{\link{delete_DB}()},
\code{\link{setup_DB}()},
\code{\link{show_DB}()}
}
\concept{DB object}
