% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_token_management.R
\name{set_redcap_token}
\alias{set_redcap_token}
\title{Sets a valid token for this session}
\usage{
set_redcap_token(DB)
}
\arguments{
\item{DB}{object generated using `load_DB()` or `setup_DB()`}
}
\value{
messages for confirmation
}
\description{
Sets a valid token for this session
}
