% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_token_management.R
\name{set_REDCap_token}
\alias{set_REDCap_token}
\title{Set a REDCap API Token}
\usage{
set_REDCap_token(DB, ask = T)
}
\arguments{
\item{ask}{Logical (TRUE/FALSE). If TRUE, asks the user for confirmation before overwriting an existing valid token. Default is `TRUE`.}
}
\value{
Invisible. A message is printed to confirm the token is successfully set.
}
\description{
Prompts the user to input a valid REDCap API token and stores it as an environment variable for the current R session.
This function validates the token before storing it. It is strongly discouraged to include API tokens directly within R scripts.
Instead of using this function you should consider setting your token within your REnviron file which can be edited with `usethis::edit_r_environ()`
}
\details{
If a valid token already exists in the R session, the function notifies the user and asks whether they want to replace it.
The user is guided to provide a new token through the console.
The token is validated internally and stored using `Sys.setenv()`.
}
\seealso{
\code{\link[RosyREDCap]{view_REDCap_token}} to view the token in your console.
\code{\link[RosyREDCap]{test_REDCap_token}} to validate the stored token.

Other token_functions: 
\code{\link{test_REDCap_token}()},
\code{\link{view_REDCap_token}()}
}
\concept{token_functions}
