% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/108_token_management.R
\name{set_REDCap_token}
\alias{set_REDCap_token}
\title{Set a REDCap API Token}
\usage{
set_REDCap_token(DB, ask = T)
}
\arguments{
\item{ask}{Logical (TRUE/FALSE). If TRUE, asks the user for confirmation before overwriting an existing valid token. Default is `TRUE`.}
}
\value{
Invisible. A message is printed to confirm the token is successfully set.
}
\description{
Prompts the user to input a valid REDCap API token and stores it as an environment variable for the current R session.
This function validates the token before storing it. It is strongly discouraged to include API tokens directly within R scripts.
Instead of using this function you should consider setting your token within your REnviron file which can be edited with `usethis::edit_r_environ()`
}
\details{
If a valid token already exists in the R session, the function notifies the user and asks whether they want to replace it.
The user is guided to provide a new token through the console.
The token is validated internally and stored using `Sys.setenv()`.
For more details, see the [pkgdown article on tokens]("../articles/Tokens.html").
For more details, see the [pkgdown article on tokens](../articles/Tokens.html).
For more details, see the [pkgdown article on tokens](vignette("Tokens")).
For more details, see the [pkgdown article on tokens](https://brandonerose.github.io/RosyREDCap/articles/Tokens.html).
For more details, see the \code{\link{setup_DB}}.
For more details, see the \code{\link{vignette("Tokens")}}.
For more details, see the \code{\url{vignette("Tokens")}}.
}
\seealso{
For more details, see the [pkgdown article on tokens](../articles/Tokens.html).
For more details, see the `vignette("Tokens")`.
For more details, see the \code{\link{vignette("Tokens")}}.
For more details, see the \code{vignette("Tokens")}.

Other Token Functions: 
\code{\link{test_REDCap_token}()},
\code{\link{view_REDCap_token}()}
}
\concept{Token Functions}
\keyword{Functions}
\keyword{Token}
