% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/108_token_management.R
\name{set_REDCap_token}
\alias{set_REDCap_token}
\title{Set a REDCap API Token to Your Current R Session}
\usage{
set_REDCap_token(DB, ask = T)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}

\item{ask}{Logical (TRUE/FALSE). If TRUE, asks the user for confirmation before overwriting an existing valid token. Default is `TRUE`.}
}
\value{
Invisible. A message is printed to confirm the token is successfully set.
}
\description{
Prompts the user to input a valid REDCap API token and stores it as an environment variable for the current R session.
Instead of using this function you should consider setting your token within your REnviron file which can be edited with \code{\link[usethis]{edit_r_environ}}.
}
\details{
If a valid token already exists in the R session, the function notifies the user and asks whether they want to replace it.
The user is guided to provide a new token through the console.
It is strongly discouraged to include API tokens directly within R scripts.
The token is validated internally and stored using `Sys.setenv()`.
}
\seealso{
For the function to work you need to have a valid DB object from \code{\link{setup_DB}()}.
See our \href{https://brandonerose.github.io/RosyREDCap/articles/Tokens.html}{REDCap Tokens Article}

Other Token Functions: 
\code{\link{test_REDCap_token}()},
\code{\link{view_REDCap_token}()}
}
\concept{Token Functions}
\keyword{Functions}
\keyword{Token}
