% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_setup_DB.R
\name{setup_DB}
\alias{setup_DB}
\title{Setup RosyREDCap}
\usage{
setup_DB(
  short_name,
  dir_path,
  redcap_base,
  force = F,
  merge_form_name = "merged",
  use_csv = F,
  auto_check_token = T
)
}
\arguments{
\item{short_name}{A character string representing the short name for the REDCap project.}

\item{dir_path}{Optional character string representing the directory path where you want the REDCap project data to be stored. If missing, DB object will only be in current R session.}

\item{redcap_base}{A character string representing the base URL of the REDCap server.}

\item{force}{Logical (TRUE/FALSE). If TRUE, forces the setup even if the `DB` object already exists. Default is `FALSE`.}

\item{merge_form_name}{A character string representing the name of the merged form. Default is "merged".}

\item{use_csv}{Logical (TRUE/FALSE). If TRUE, uses CSV files for data storage. Default is `FALSE`.}

\item{auto_check_token}{Logical (TRUE/FALSE). If TRUE, automatically checks the validity of the REDCap API token. Default is `TRUE`.}
}
\value{
The `DB` object with the REDCap API token and configurations set.
}
\description{
Initializes the `DB` object with the necessary REDCap API token and other configurations.
}
\details{
This function sets up the `DB` object by storing the REDCap API token and other configurations required for interacting with the REDCap server.
It ensures that the token is valid and ready for use in subsequent API calls.
}
\examples{
\dontrun{
# Initialize the DB object with the REDCap API token and URL
DB <- setup_DB(
  short_name = "ABC",
  dir_path = "path/to/secure/file/storage",
  redcap_base = "https://redcap.yourinstitution.edu/"
)
}
}
\seealso{
Other DB object: 
\code{\link{delete_DB}()},
\code{\link{load_DB}()},
\code{\link{save_DB}()},
\code{\link{show_DB}()}
}
\concept{DB object}
