% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_DB.R
\name{setup_DB}
\alias{setup_DB}
\title{Setup for DB including token}
\usage{
setup_DB(
  short_name,
  dir_path,
  token_name,
  redcap_base_link,
  force = F,
  merge_form_name,
  validate = T,
  use_csv = F
)
}
\arguments{
\item{short_name}{character name as a shortcut}

\item{dir_path}{character file path of the directory}

\item{token_name}{character string of what the token is called when using Sys.setenv and Sys.getenv}

\item{redcap_base_link}{character of the base REDCap link, ex. https://redcap.miami.edu}

\item{force}{logical for force blank load vs last save}

\item{merge_form_name}{name of merged non-repeating to be used in package}

\item{validate}{logical for validation}
}
\value{
DB
}
\description{
Setup for DB including token
}
