% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_DB_mod.R
\name{setup_RosyREDCap}
\alias{setup_RosyREDCap}
\title{Setup for DB including token}
\usage{
setup_RosyREDCap(
  short_name,
  dir_path,
  token_name,
  redcap_base,
  force = F,
  merge_form_name = "merged",
  validate = T,
  use_csv = F
)
}
\arguments{
\item{short_name}{character name as a shortcut}

\item{dir_path}{character file path of the directory}

\item{token_name}{character string of what the token is called when using Sys.setenv and Sys.getenv}

\item{redcap_base}{character of the base REDCap link, ex. https://redcap.miami.edu}

\item{force}{logical for force blank load vs last save}

\item{merge_form_name}{name of merged non-repeating to be used in package}

\item{validate}{logical for validation check}

\item{DB}{a list object that contains all metadata, data, log, transformations, and more for one REDCap project.}
}
\value{
DB
}
\description{
Setup for DB including token
}
