% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_setup_DB.R
\name{show_DB}
\alias{show_DB}
\title{Shows DB in the env}
\usage{
show_DB(DB, also_metadata = T, only_dfs = T)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \code{DB <- \link{load_DB}("PROJ")} or \link{setup_DB}()}

\item{also_metadata}{logical for including metadata}

\item{only_dfs}{logical for including data.frames}

\item{data_choice}{whether to use 'data' or 'data'}
}
\value{
DB tables
}
\description{
Shows DB in the env
}
\seealso{
Other DB object: 
\code{\link{delete_DB}()},
\code{\link{save_DB}()},
\code{\link{setup_DB}()}
}
\concept{DB object}
