% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_db.R
\name{show_DB}
\alias{show_DB}
\title{Shows DB in the env}
\usage{
show_DB(DB)
}
\arguments{
\item{DB}{list for the package that contains applications and reference files}
}
\value{
DB tables in your environment
}
\description{
Shows DB in the env
}
