% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/116_summarize_DB.R
\name{summarize_DB}
\alias{summarize_DB}
\title{summarize_DB}
\usage{
summarize_DB(
  DB,
  with_links = T,
  deidentify = T,
  clean = T,
  drop_blanks = T,
  other_drops = NULL,
  include_metadata = T,
  annotate_metadata = T,
  include_record_summary = T,
  include_users = T,
  include_log = T,
  separate = F,
  force = F
)
}
\arguments{
\item{drop_blanks}{optional logical for dropping blanks}
}
\description{
summarize_DB
}
