% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_transform.R
\name{transform_DB}
\alias{transform_DB}
\title{Transform DB}
\usage{
transform_DB(
  DB,
  merge_non_rep_to_reps = F,
  records = NULL,
  force = F,
  terminal_transformation = F
)
}
\arguments{
\item{merge_non_rep_to_reps}{logical for merging non-repeating to every repeating instrument}

\item{records}{character vector of records to be included in transform}

\item{force}{logical for merging forcing transoformation even if up to date}

\item{terminal_transformation}{logical for whether this transformation is terminal/destructive}
}
\value{
DB object that has DB$data_transform, can be based on a remap file from input folder or default
}
\description{
Transform DB
}
