% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_update_DB.R
\name{update_DB}
\alias{update_DB}
\title{Shows DB in the env}
\usage{
update_DB(
  DB,
  force = F,
  day_of_log = 10,
  clean = T,
  get_files = T,
  original_file_names = F
)
}
\arguments{
\item{DB}{DB from load_DB or setup_DB}

\item{force}{logical for force a fresh update}

\item{day_of_log}{numbers of days to be checked in the log}

\item{clean}{logical for wether or not to return raw or clean REDCap. Default is TRUE.}

\item{get_files}{logical for whether or not to get files from redcap.}

\item{original_file_names}{logical for whether or not to use original file names.}
}
\value{
messages for confirmation
}
\description{
Shows DB in the env
}
