% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_update_DB.R
\name{update_DB}
\alias{update_DB}
\title{Shows DB in the env}
\usage{
update_DB(token, force = F, day_of_log = 10, use_missing_codes = T)
}
\arguments{
\item{token}{REDCap API token}

\item{force}{logical for force a fresh update}
}
\value{
messages for confirmation
}
\description{
Shows DB in the env
}
