% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_transform.R
\name{upload_transform_to_DB}
\alias{upload_transform_to_DB}
\title{upload_transform_to_DB Transform}
\usage{
upload_transform_to_DB(DB)
}
\arguments{
\item{DB}{a list object slightly modified version of RosyDB and contains all metadata, data, transformations, and more}
}
\description{
upload_transform_to_DB Transform
}
