% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/115_transform.R
\name{upload_transform_to_DB}
\alias{upload_transform_to_DB}
\title{upload_transform_to_DB Transform}
\usage{
upload_transform_to_DB(DB)
}
\arguments{
\item{DB}{A validated `DB` object containing REDCap project data and settings. Generated using \link{load_DB} or \link{setup_DB}}
}
\description{
upload_transform_to_DB Transform
}
