% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_token_management.R
\name{view_REDCap_token}
\alias{view_REDCap_token}
\title{View the REDCap API token currently stored in the session}
\usage{
view_REDCap_token(DB)
}
\arguments{
\item{DB}{a list object that contains all metadata, data, log, transformations, and more for one REDCap project.}
}
\value{
REDCap API token currently stored in the session
}
\description{
View the REDCap API token currently stored in the session
}
