% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_token_management.R
\name{view_redcap_token}
\alias{view_redcap_token}
\title{View the REDCap API token currently stored in the session}
\usage{
view_redcap_token(DB)
}
\value{
REDCap API token currently stored in the session
}
\description{
View the REDCap API token currently stored in the session
}
